xcdr test: fixed warnings and pre-C++11 compat

+++ tests/DCPS/Compiler/xcdr/xcdr.cpp
@@ -1603,13 +1603,13 @@ void expect_values_equal(const MixedMutableStruct& a,
 {
   expect_values_equal(a.struct_nested, b.struct_nested);
   EXPECT_EQ(a.sequence_field.length(), b.sequence_field.length());
-  for (size_t i = 0; i < a.sequence_field.length(); ++i) {
+  for (unsigned i = 0; i < a.sequence_field.length(); ++i) {
     EXPECT_EQ(a.sequence_field[i], b.sequence_field[i]);
   }
   EXPECT_EQ(a.union_nested._d(), b.union_nested._d());
   EXPECT_STREQ(a.union_nested.string_field(), b.union_nested.string_field());
   EXPECT_EQ(a.sequence_field2.length(), b.sequence_field2.length());
-  for (size_t i = 0; i < a.sequence_field2.length(); ++i) {
+  for (unsigned i = 0; i < a.sequence_field2.length(); ++i) {
     EXPECT_STREQ(a.sequence_field2[i], b.sequence_field2[i]);
   }
 }
@@ -1700,7 +1700,7 @@ void expect_values_equal(const NestingFinalStruct& a,
   EXPECT_STREQ(a.string_field, b.string_field);
   expect_values_equal_base(a.appendable_nested, b.appendable_nested);
   EXPECT_EQ(a.sequence_field.length(), b.sequence_field.length());
-  for (size_t i = 0; i < a.sequence_field.length(); ++i) {
+  for (unsigned i = 0; i < a.sequence_field.length(); ++i) {
     EXPECT_EQ(a.sequence_field[i], b.sequence_field[i]);
   }
   expect_values_equal_base(a.mutable_nested, b.mutable_nested);
@@ -1751,7 +1751,7 @@ void set_values(NestingAppendableStruct& value)
   value.string_field = "hello world";
   set_values(value.mutable_nested);
   value.sequence_field.length(3);
-  for (size_t i = 0; i < value.sequence_field.length(); ++i) {
+  for (unsigned i = 0; i < value.sequence_field.length(); ++i) {
     value.sequence_field[i] = 0x7fffffff;
   }
   set_values(value.final_nested);
@@ -1764,7 +1764,7 @@ void expect_values_equal(const NestingAppendableStruct& a,
   EXPECT_STREQ(a.string_field, b.string_field);
   expect_values_equal_base(a.mutable_nested, b.mutable_nested);
   EXPECT_EQ(a.sequence_field.length(), b.sequence_field.length());
-  for (size_t i = 0; i < a.sequence_field.length(); ++i) {
+  for (unsigned i = 0; i < a.sequence_field.length(); ++i) {
     EXPECT_EQ(a.sequence_field[i], b.sequence_field[i]);
   }
   expect_values_equal_base(a.final_nested, b.final_nested);
@@ -1815,7 +1815,7 @@ void set_values(NestingMutableStruct& value)
   value.string_field = "hello world";
   set_values(value.appendable_nested);
   value.sequence_field.length(7);
-  for (size_t i = 0; i < value.sequence_field.length(); ++i) {
+  for (unsigned i = 0; i < value.sequence_field.length(); ++i) {
     value.sequence_field[i] = 0x7f;
   }
   set_values(value.final_nested);
@@ -1828,7 +1828,7 @@ void expect_values_equal(const NestingMutableStruct& a,
   EXPECT_STREQ(a.string_field, b.string_field);
   expect_values_equal_base(a.appendable_nested, b.appendable_nested);
   EXPECT_EQ(a.sequence_field.length(), b.sequence_field.length());
-  for (size_t i = 0; i < a.sequence_field.length(); ++i) {
+  for (unsigned i = 0; i < a.sequence_field.length(); ++i) {
     EXPECT_EQ(a.sequence_field[i], b.sequence_field[i]);
   }
   expect_values_equal_base(a.final_nested, b.final_nested);
@@ -1900,7 +1900,7 @@ void key_only_test()
   set_values(value);
   Type result;
   NestedKeyOnly<Type> wrapped_result(result);
-  amalgam_serializer_test_base<NestedKeyOnly<const Type>, NestedKeyOnly<Type>>(
+  amalgam_serializer_test_base<NestedKeyOnly<const Type>, NestedKeyOnly<Type> >(
     xcdr1, expected, wrapped_value, wrapped_result);
   EXPECT_PRED_FORMAT2(assert_values, wrapped_value, wrapped_result);
 }
